--- src/adapters/public-api.js
@@ -1,9 +1,0 @@
-// import * as hapoalim from './hapoalim/public-api';
-// import * as leumi from './leumi/public-api';
-export * from './visa-cal/public-api';
-
-import * as puppeteerAdapters from './puppeteer';
-import runner from './runner';
-import * as session from './session';
-
- export { puppeteerAdapters, runner, session };
