+++ src/hapoalim/adapters/login.js
@@ -1,6 +1,6 @@
 import { fillInputs, clickButton, waitUntilElementFound } from '../../helpers/elements-interactions';
 import { navigateTo, getCurrentUrl, waitForRedirect } from '../../helpers/navigation';
-import { BASE_URL, LOGIN_URL } from './definitions';
+import { BASE_URL, LOGIN_URL } from '../definitions';
 import getKeyByValue from '../../helpers/filters';
 import { SCRAPE_PROGRESS_TYPES, LOGIN_RESULT } from '../../constants';
 import { handleLoginResult, isValidCredentials } from '../../helpers/login';
@@ -27,7 +27,7 @@ function getPossibleLoginResults() {
   return urls;
 }
 
-function loginAdapter(options) {
+export function loginAdapter(options) {
   return {
     name: 'scrapeTransactions(hapoalim)',
     validate: (context) => {
@@ -61,5 +61,3 @@ function loginAdapter(options) {
     },
   };
 }
-
-export default loginAdapter;
